Write Unicode to console but native CP to file
Better fix for issues #11 and #14. Uses the same basic algorithm as the native VC++ compiler so that writing to the console directly will set the mode to Unicode (with BOM); however, redirecting to a file (which various operations in both cmd.exe and powershell.exe do) will use the native console codepage.