Colorize vswhere output (#244)
* Refactor formatters to consolidate arguments

* Colorize output by default unless piped

* Opt into color and add tests

Fails in PowerShell if you do something like `[xml](vswhere -format xml)` if color is enabled, so had to add -color switch. Not ideal since it otherwise disables color when piped, but still easier to read if colorized explicitly.

* Enable colors by default, add -nocolor switch

* Resolve PR feedback

* Ensure console is initialized

Could simply call Initialize wherever it needs to be called, but would rather be explicit.